+++ apps/web-antd/src/views/system/dict/data/index.vue
@@ -128,9 +128,14 @@ onMounted(() => {
     dictType.value = value;
     await tableApi.query();
   });
+  emitter.on('reset', async () => {
+    dictType.value = '';
+    await tableApi.reload();
+  });
 });
 onBeforeUnmount(() => {
   emitter.off('rowClick');
+  emitter.off('reset');
 });
 </script>
 
@@ -174,7 +179,9 @@ onBeforeUnmount(() => {
           >
             {{ $t('pages.common.edit') }}
           </action-button>
+          <!-- 这里数据会不一致 必须加key标识 -->
           <Popconfirm
+            :key="row.dictCode"
             placement="left"
             title="确认删除？"
             @confirm="handleDelete(row)"