+++ apps/web-antd/src/views/system/dict/mitt.ts
@@ -4,6 +4,7 @@ import { mitt } from '@vben/utils';
  * dictType: string
  */
 type Events = {
+  reset: void;
   rowClick: string;
 };