+++ packages/@core/ui-kit/form-ui/src/form-render/form-label.vue
@@ -1,7 +1,11 @@
 <script setup lang="ts">
 import type { CustomRenderType } from '../types';
 
-import { FormLabel, VbenHelpTooltip } from '@vben-core/shadcn-ui';
+import {
+  FormLabel,
+  VbenHelpTooltip,
+  VbenRenderContent,
+} from '@vben-core/shadcn-ui';
 import { cn } from '@vben-core/shared/utils';
 
 interface Props {
@@ -20,12 +24,9 @@ const props = defineProps<Props>();
     <span v-if="required" class="mr-0.5 text-destructive">*</span>
     <slot></slot>
     <VbenHelpTooltip v-if="help" trigger-class="size-3.5 ml-1">
-      <!-- 可通过\n换行 -->
-      <span class="whitespace-pre-line">
-        {{ help }}
-      </span>
-      <!-- <VbenRenderContent :content="help" /> -->
+      <VbenRenderContent :content="help" render-br />
     </VbenHelpTooltip>
+    <slot name="extra"></slot>
     <span v-if="colon && label" class="ml-0.5">:</span>
   </FormLabel>
 </template>